Implemented SafeWalk
