+++ BleachHack-Fabric-1.15/src/main/java/bleach/hack/module/ModuleManager.java
@@ -72,6 +72,7 @@ public class ModuleManager {
 			new PacketFly(),
 			new Peek(),
 			new PlayerCrash(),
+			new SafeWalk(),
 			new Scaffold(),
 			new Spammer(),
 			new SpeedHack(),
