+++ BleachHack-Fabric-1.15/src/main/java/bleach/hack/module/mods/SafeWalk.java
@@ -1,0 +1,10 @@
+package bleach.hack.module.mods;
+
+import bleach.hack.module.Category;
+import bleach.hack.module.Module;
+
+public class SafeWalk extends Module {
+    public SafeWalk() {
+        super("SafeWalk", KEY_UNBOUND, Category.MOVEMENT, "Stops you walking off blocks");
+    }
+}
