+++ BleachHack-Fabric-1.15/src/main/resources/bleachhack.mixins.json
@@ -19,7 +19,7 @@
     "MixinLlamaEntity",
     "MixinMinecraftClient",
     "MixinPacketInflater",
-    "MixinPlayerEntity",
+    "MixinClientPlayerEntity",
     "MixinPlayerInventory",
     "MixinRenderTickCounter",
     "MixinScreen",
